refactor: new type of message processing result for resolving circular dependency over PeerManager
